Revert "BoardConfig: do not use memory tweaks"
Summary: This reverts commit f0aa700e1ce7e96683c3a1e73abdc022cdaf4289. I've noticed as have others that after removing the memory values we were using, the device would slow incredibly after being on for awhile, even becoming unstable. I think we should modify the values ourselves in the low ram init instead of removing them entirely. This is strange because there's no real way to show numbers for or against, but the popular opinion is that this commit only made things worse.

Test Plan: Left device on for an hour, tested that it retained the same speed as it had 5 minutes from powering.

Differential Revision: http://review.codefirex.com/D102